Refactor splitting main_mode module
